// frontend/src/app/Models/user.ts
export interface User {
    id?: number; 
    email: string;
    password?: string; 
    name: string;
    role?: string;
    phone?: string;
    address?: string;
    country?: string;
    region?: string;
    codepostal?: string;
    created_at?: Date;
    updated_at?: Date;
    picturesrc?: string;
  }